Minor improvements to code modularity
The ClientWindow class has now been renamed "UIHandle" and a new
"NetworkHandle" class has been added. From now on these classes will be
designed to handle GUI and Networking related actions respectively.

Separated all the code that use to be in the server file into its own
separate files within the "shared" package. These files are the
"visuals" file and the "networking" file, which are meant to be used for
handling GUI and Networking related actions that are not unique to
either the client or server.